hot fix

+++ 0x15-file_io/2-append_text_to_file.c
@@ -6,7 +6,7 @@
  *
  * Return: the length of the string
  */
-size_t _strlen(char *str)
+int _strlen(char *str)
 {
 	int length = 0;
